+++ lib/src/editor/widgets/text/utils/text_block_utils.dart
@@ -1,7 +1,10 @@
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+
 import '../../../../common/structs/horizontal_spacing.dart';
 import '../../../../document/attribute.dart';
 import '../../../../document/nodes/block.dart';
+import '../../../../document/nodes/node.dart';
 import '../../default_styles.dart';
 
 typedef LeadingBlockIndentWidth = HorizontalSpacing Function(
@@ -12,6 +15,30 @@ typedef LeadingBlockIndentWidth = HorizontalSpacing Function(
 
 typedef LeadingBlockNumberPointWidth = double Function(
     double fontSize, int count);
+
+typedef TextSpanBuilder = InlineSpan Function(
+  BuildContext context,
+  Node node,
+  int nodeOffset,
+  String text,
+  TextStyle? style,
+  GestureRecognizer? recognizer,
+);
+
+TextSpan defaultSpanBuilder(
+  BuildContext context,
+  Node node,
+  int textOffset,
+  String text,
+  TextStyle? style,
+  GestureRecognizer? recognizer,
+) =>
+    TextSpan(
+      text: text,
+      style: style,
+      recognizer: recognizer,
+      mouseCursor: (recognizer != null) ? SystemMouseCursors.click : null,
+    );
 
 abstract final class TextBlockUtils {
   /// Get the horizontalSpacing using the default